Add IDxcLangExtensions2 to set target triple. (#2981)
* Add IDxcLangExtensions2 to set target triple.
